be more deliberate about the accounts we manage

+++ entrypoint.sh
@@ -12,20 +12,33 @@ if [ -n "$INIT_SQL" ]; then
   # semicolons (no line breaks or comments are permitted).
   # TODO proper SQL escaping on ALL the things D:
 
+  if [ -z "$MYSQL_ROOT_PASSWORD" -a -z "$MYSQL_ALLOW_EMPTY_PASSWORD" ]; then
+    echo >&2 'error: database is uninitialized and MYSQL_ROOT_PASSWORD not set'
+    echo >&2 '  Did you forget to add -e MYSQL_ROOT_PASSWORD=... ?'
+    exit 1
+  fi
+
   rm -f "$INIT_SQL"
 
   # Basics
-  echo "DELETE FROM mysql.user;" >> "$INIT_SQL"
+
+  # Setup root
+  echo "DROP USER IF EXISTS 'root'@'%';" >> "$INIT_SQL"
   echo 'FLUSH PRIVILEGES ;' >> "$INIT_SQL"
   echo "CREATE USER 'root'@'%' IDENTIFIED BY '${MYSQL_ROOT_PASSWORD}';" >> "$INIT_SQL"
   echo "GRANT ALL ON *.* TO 'root'@'%' WITH GRANT OPTION;" >> "$INIT_SQL"
+
+  # Delete extraneous dbs
   echo "DROP DATABASE IF EXISTS test;" >> "$INIT_SQL"
 
+  # Create the schema
   if [ -n "$MYSQL_DATABASE" ]; then
     echo "CREATE DATABASE IF NOT EXISTS \`$MYSQL_DATABASE\` ;" >> "$INIT_SQL"
   fi
 
   if [ "$MYSQL_USER" -a "$MYSQL_PASSWORD" ]; then
+    echo "DROP USER IF EXISTS '$MYSQL_USER'@'%';" >> "$INIT_SQL"
+    echo 'FLUSH PRIVILEGES ;' >> "$INIT_SQL"
     echo "CREATE USER '$MYSQL_USER'@'%' IDENTIFIED BY '$MYSQL_PASSWORD' ;" >> "$INIT_SQL"
     
     if [ "$MYSQL_DATABASE" ]; then
@@ -43,11 +56,6 @@ if [ "$1" = 'mysqld_safe' ] || [ "$1" = 'mysqld' ]; then
 	
 	if [ ! -d "$DATADIR/mysql" ]; then
     echo "Detected fresh install"
-		if [ -z "$MYSQL_ROOT_PASSWORD" -a -z "$MYSQL_ALLOW_EMPTY_PASSWORD" ]; then
-			echo >&2 'error: database is uninitialized and MYSQL_ROOT_PASSWORD not set'
-			echo >&2 '  Did you forget to add -e MYSQL_ROOT_PASSWORD=... ?'
-			exit 1
-		fi
 		
 		echo 'Running mysql_install_db ...'
 		time mysql_install_db --datadir="$DATADIR"
